Moving concrete implementations of the element face accessors to reduce number of headers. No change in number of translation units
